requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.0.2
click==8.1.7
colorama==0.4.6
coverage==7.4.1
dill==0.3.8
exceptiongroup==1.2.0
Faker==23.1.0
fastapi==0.109.2
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
markdown2==2.4.13
mccabe==0.7.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pylint==3.0.3
pylint-pytest==1.1.7
pytest==8.0.0
pytest-cov==4.1.0
pytest-pylint==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
six==1.16.0
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
